normalized awful type mess, but even more bugs. weird crashes, etc

+++ pk/pk.h
@@ -106,7 +106,7 @@ typedef struct {
 } due_cacheline_t;
       
 typedef void (*trap_handler)(trapframe_t*); //MWG
-typedef int (*user_due_trap_handler)(trapframe_t*, float_trapframe_t*, long, due_candidates_t*, due_cacheline_t*, word_t*, short, short, short, short, short); //MWG
+typedef int (*user_due_trap_handler)(trapframe_t*, float_trapframe_t*, long, due_candidates_t*, due_cacheline_t*, word_t*, size_t, size_t, int, int, int); //MWG
 int default_memory_due_trap_handler(trapframe_t*, int error_code, const char* expl); //MWG
 void sys_register_user_memory_due_trap_handler(user_due_trap_handler fptr); //MWG
 
@@ -123,16 +123,16 @@ int copy_trapframe(trapframe_t* dest, trapframe_t* src); //MWG
 int copy_float_trapframe(float_trapframe_t* dest, float_trapframe_t* src); //MWG
 long decode_load_vaddr(long insn, trapframe_t* tf); //MWG
 long decode_i_imm(long insn); //MWG
-unsigned decode_rs1(long insn); //MWG
-unsigned decode_rd(long insn); //MWG
-short decode_regfile(long insn); //MWG
-int load_value_from_message(word_t* recovered_message, word_t* load_value, due_cacheline_t* cl, unsigned load_size, int offset); //MWG
-int writeback_recovered_message(word_t* recovered_message, word_t* load_value, trapframe_t* tf, short mem_type, unsigned rd, short float_regfile); //MWG 
-int get_float_register(unsigned frd, unsigned long* raw_value); //MWG
-int set_float_register(unsigned frd, unsigned long raw_value); //MWG
+size_t decode_rs1(long insn); //MWG
+size_t decode_rd(long insn); //MWG
+int decode_regfile(long insn); //MWG
+int load_value_from_message(word_t* recovered_message, word_t* load_value, due_cacheline_t* cl, size_t load_size, int offset); //MWG
+int writeback_recovered_message(word_t* recovered_message, word_t* load_value, trapframe_t* tf, int mem_type, size_t rd, int float_regfile); //MWG 
+int get_float_register(size_t frd, unsigned long* raw_value); //MWG
+int set_float_register(size_t frd, unsigned long raw_value); //MWG
 int set_float_trapframe(float_trapframe_t* float_tf); //MWG
 void dump_word(word_t* w); //MWG
-int compare_recovery(word_t* recovered_value, word_t* cheat_msg, word_t* recovered_load_value, word_t* cheat_load_value, short demand_load_message_offset); //MWG
+int compare_recovery(word_t* recovered_value, word_t* cheat_msg, word_t* recovered_load_value, word_t* cheat_load_value, int demand_load_message_offset); //MWG
 
 typedef struct {
   int elf64;
